Refactor code + split maze gen from draw()
Currently the maze generates instantly outside of the draw() loop. Doing this changed the functionality of how the maze is generated since it's no longer using draw() as the while loop. Will be updated to allow the user to use both if they want to visualize the generation or if they want instant generation of the maze.